MAINT move datasets.py to examples/
