+++ examples/1d/plot_classif_torch.py
@@ -55,7 +55,7 @@ import matplotlib.pyplot as plt
 # `fetch_fsdd` function downloads the FSDD, if needed.
 
 from kymatio.torch import Scattering1D
-from kymatio.datasets import fetch_fsdd
+from ...datasets import fetch_fsdd
 
 ###############################################################################
 # Pipeline setup
